weather_monitoring_system/requirements.txt
requests
psycopg2-binary  # For PostgreSQL database interaction
plotly            # For data visualization 
